+++ packages/web3-utils/CHANGELOG.md
@@ -221,7 +221,9 @@ Documentation:
 ### Added
 
 - `toWei` add warning when using large numbers or large decimals that may cause precision loss (#6908)
+- `toWei` and `fromWei` now supports integers as a unit. (#7053)  
 
 ### Fixed
 
 - `toWei` support numbers in scientific notation (#6908)
+- `toWei` and `fromWei` now supports integers as a unit. (#7053)